refactor: update import paths for helper functions and replace camelCase with kebab-case
