+++ main-app/src/app/api/(auth)/email-verification/route.ts
@@ -1,4 +1,4 @@
-import { generateTokenPair } from '@/helpers/jwtToken';
+import { generateTokenPair } from '@/helpers/jwt-token';
 import { sendResponse } from '@/lib/apis/send-response';
 import { createUserSession } from '@/lib/auth/session-tracker';
 import dbConnect from '@/lib/config/db-connect';
