+++ main-app/src/helpers/generate-otp.ts
