+++ main-app/src/helpers/hash-password.ts
