+++ main-app/src/lib/auth/auth-middleware.ts
@@ -1,4 +1,4 @@
-import { authenticateToken } from '@/helpers/jwtToken';
+import { authenticateToken } from '@/helpers/jwt-token';
 import { sendResponse } from '@/lib/apis/send-response';
 import { touchSession } from '@/lib/auth/session-tracker';
 import dbConnect from '@/lib/config/db-connect';
